requirements.txt
news_2_pdf>=0.0.11
python-telegram-bot
pyyaml
telegram_util>=0.0.28
channel2pdf>=0.0.2